@rkistner 2024-08-14
Will get a new release with these changes out next week.

Note that for the web SDK, the changes in https://github.com/powersync-ja/powersync-js/pull/266 fix the biggest performance issues, but these changes will give further improvements for buckets with many more operations than actual rows.